V2 controller metrics concurrency fix (#231)
* Refine the metrics status


* Refine the metrics status


* Fix test case error


* Use group instead of ASG


* Ignore generated code


* Ignore generated code


* Fix the concurrent issue


* Fix the concurrent issue


* Move metrics related functions into RollingUpgradeContext


* Move metrics related functions into RollingUpgradeContext


* Move metrics related functions into upgrade_metrics.go


* Move metrics related functions into metrics.go
